+++ poms/configuration_sharing/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ConfigurationSharingConfig(AppConfig):
+    name = 'poms.configuration_sharing'