+++ poms/ui/serializers.py
@@ -52,7 +52,7 @@ class TemplateListLayoutSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = TemplateListLayout
-        fields = ['id', 'member', 'content_type', 'name', 'is_default', 'data']
+        fields = ['id', 'member', 'content_type', 'name', 'is_default', 'data',]
 
 
 class TemplateEditLayoutSerializer(serializers.ModelSerializer):
@@ -71,7 +71,7 @@ class ContextMenuLayoutSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = ContextMenuLayout
-        fields = ['id', 'member', 'type', 'name', 'data']
+        fields = ['id', 'member', 'type', 'name', 'data', 'origin_for_global_layout', 'sourced_from_global_layout']
 
 
 class ListLayoutSerializer(serializers.ModelSerializer):
@@ -81,7 +81,7 @@ class ListLayoutSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = ListLayout
-        fields = ['id', 'member', 'content_type', 'name', 'is_default', 'is_active', 'data']
+        fields = ['id', 'member', 'content_type', 'name', 'is_default', 'is_active', 'data', 'origin_for_global_layout', 'sourced_from_global_layout']
 
 
 class DashboardLayoutSerializer(serializers.ModelSerializer):
@@ -90,7 +90,7 @@ class DashboardLayoutSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = DashboardLayout
-        fields = ['id', 'member', 'name', 'is_default', 'is_active', 'data']
+        fields = ['id', 'member', 'name', 'is_default', 'is_active', 'data', 'origin_for_global_layout', 'sourced_from_global_layout']
 
 
 class ConfigurationExportLayoutSerializer(serializers.ModelSerializer):
@@ -109,7 +109,7 @@ class EditLayoutSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = EditLayout
-        fields = ['id', 'member', 'content_type', 'data']
+        fields = ['id', 'member', 'content_type', 'data', 'origin_for_global_layout', 'sourced_from_global_layout']
 
 
 class BookmarkRecursiveField(serializers.Serializer):