+++ poms_app/settings.py
@@ -119,6 +119,7 @@ INSTALLED_APPS = [
 
     'poms.reconciliation',
     'poms.file_reports',
+    'poms.configuration_sharing',
 
     'django.contrib.admin',
     'django.contrib.admindocs',