+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 21.01.2020
+- Layout Sharing
+- Import Configuration
+
 ## 12.12.2019
 - Reconciliation update